src/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
databases==0.9.0
python-dotenv==1.1.1
fuzzywuzzy[speedup]==0.18.0
pytz==2025.2
pytest==8.4.2
httpx==0.28.1
pytest-asyncio==1.2.0
slowapi==0.1.9
coverage==7.10.6